Split DeltaSharingSourceSuite
